requirements.txt
tiktoken 
chromadb
langchain
flask
GitPython
python-dotenv
langchain-community
langchain-groq
groq
sentence-transformers